requirements.txt
Flask
Flask-Login
supabase
python-dotenv
gotrue
gunicorn